Centralize routes, and add actions
Centralize the definition of routes to an `illithid.routes` namespace
with the existing data for title and component for each existing route,
and add an optional :action for routes, which displays a button in the
upper right hand corner of the screen. Use this option to turn the
"create new character" button into a Plus icon on the navbar for the
character screen.