Fix resolution of specs for resources
Because the spec resolution is running during the compilation of the
Clojurescript version of the resource file, the macro runtime that's
running the splice operation hasn't actually gotten a chance to require
the namespace containing the definition of the spec for the resource.

This is a rough attempt at trying to require the namespace for any
specs passed in to the `defresources` macro, falling back on just
printing a warning to stderr if it's not possible.

Just in case it *isn't* possible, this also adds a simple test for the
spells namespace to make sure all spells pass the spec, and adds the
appropriate entry to circle.yml to run that test